+++ src/features/recruitment/ui/modal/index.tsx
@@ -2,10 +2,10 @@ import { useState } from 'react'
 
 import { useRecruitmentTagsQuery } from '@/api/recruitment/useRecruitmentTagsQuery'
 import Modal from '@/components/common/Modal'
-import SelectedTagList from '@/components/common/recruitment/modal/SelectedTagList'
-import TagFilterActionButtons from '@/components/common/recruitment/modal/TagFilterActionButtons'
-import TagOptionList from '@/components/common/recruitment/modal/TagOptionList'
-import TagSearchInput from '@/components/common/recruitment/modal/TagSearchInput'
+import SelectedTagList from '@/features/recruitment/ui/modal/SelectedTagList'
+import TagFilterActionButtons from '@/features/recruitment/ui/modal/TagFilterActionButtons'
+import TagOptionList from '@/features/recruitment/ui/modal/TagOptionList'
+import TagSearchInput from '@/features/recruitment/ui/modal/TagSearchInput'
 import { useRecruitmentModalStore } from '@/store/recruitment/useRecruitmentModalStore'
 import { useRecruitmentTagsStore } from '@/store/recruitment/useRecruitmentTagsStore'
